Merge develop into main

+++ IO.SDK/DataPoolMonitoring.cpp
@@ -1,3 +1,13 @@
+/**
+ * @file DataPoolMonitoring.cpp
+ * @author Sylvain Guillet (sylvain.guillet@live.com)
+ * @brief 
+ * @version 0.1
+ * @date 2021-07-03
+ * 
+ * @copyright Copyright (c) 2021
+ * 
+ */
 #include "DataPoolMonitoring.h"
 
 IO::SDK::DataPoolMonitoring IO::SDK::DataPoolMonitoring::m_instance;
@@ -19,7 +29,7 @@ std::vector<std::string> IO::SDK::DataPoolMonitoring::GetStringProperty(const st
 
 	if (found)
 	{
-		for (size_t i = 0; i < n; i++)
+		for (int i = 0; i < n; i++)
 		{
 			res.push_back(values[i]);
 		}
@@ -38,7 +48,7 @@ std::vector<int> IO::SDK::DataPoolMonitoring::GetIntegerProperty(const std::stri
 
 	std::vector<int> res;
 
-	for (size_t i = 0; i < n; i++)
+	for (int i = 0; i < n; i++)
 	{
 		res.push_back(values[i]);
 	}
@@ -57,7 +67,7 @@ std::vector<double> IO::SDK::DataPoolMonitoring::GetDoubleProperty(const std::st
 
 	std::vector<double> res;
 
-	for (size_t i = 0; i < n; i++)
+	for (int i = 0; i < n; i++)
 	{
 		res.push_back(values[i]);
 	}
